feat: added KeyValueStore.layerStorage
