+++ .changeset/fifty-chairs-jam.md
@@ -1,0 +1,5 @@
+---
+"@effect/platform": patch
+---
+
+added KeyValueStore.layerStorage to wrap instances of the `Storage` type.
